+++ src/App.tsx
@@ -12,10 +12,6 @@ import {
 } from "./components/widgets/commitButton/CommitButton";
 import { Finance, financePath } from "./components/widgets/finance/Finance";
 import { indexPath } from "./components/widgets/path";
-import {
-  PurchaseOverview,
-  purchaseOverviewPath
-} from "./components/widgets/purchaseOverview/PurchaseOverview";
 import { Redeem, redeemPath } from "./components/widgets/redeem/Redeem";
 import { GlobalStyle } from "./globalStyles";
 
@@ -30,10 +26,6 @@ function App() {
           <Route path={financePath} element={<Finance />}></Route>
           <Route path={commitPath} element={<Commit />}></Route>
           <Route path={commitButtonPath} element={<CommitButton />}></Route>
-          <Route
-            path={purchaseOverviewPath}
-            element={<PurchaseOverview />}
-          ></Route>
           <Route path={redeemPath} element={<Redeem />}></Route>
         </Routes>
       </HashRouter>
