+++ src/components/widgets/redeem/Redeem.tsx
@@ -4,10 +4,16 @@ import {
   RedemptionWidgetAction,
   subgraph
 } from "@bosonprotocol/react-kit";
-import { DeliveryInfoCallbackResponse } from "@bosonprotocol/react-kit/dist/cjs/hooks/callbacks/types";
 import { useSearchParams } from "react-router-dom";
 
 import { CONFIG, getMetaTxConfig } from "../../../config";
+import { extractBooleanParam } from "../../../utils/params";
+import {
+  createDeliveryInfoHandler,
+  createRedemptionConfirmedHandler,
+  createRedemptionSubmittedHandler,
+  parseDeliveryInfoData
+} from "../../../utils/redeem";
 
 export const redeemPath = "/redeem";
 export function Redeem() {
@@ -23,76 +29,20 @@ export function Redeem() {
   const exchangeState: subgraph.ExchangeState = checkExchangeState(
     searchParams.get("exchangeState") || undefined
   );
-  const deliveryInfo = searchParams.get("deliveryInfo") || undefined;
-  let deliveryInfoDecoded = undefined;
-  if (deliveryInfo) {
-    try {
-      deliveryInfoDecoded = JSON.parse(deliveryInfo);
-      for (const key of Object.keys(deliveryInfoDecoded)) {
-        deliveryInfoDecoded[key] = decodeURIComponent(deliveryInfoDecoded[key]);
-      }
-    } catch (e) {
-      console.error(
-        `Unable to parse JSON from deliveryInfo='${deliveryInfo}': ${e}`
-      );
-    }
-  }
-  const sendDeliveryInfoThroughXMTP = extractBooleanParam(
-    searchParams.get("sendDeliveryInfoThroughXMTP"),
-    true
-  );
-  const targetOrigin = searchParams.get("targetOrigin") || undefined;
-  const shouldWaitForResponse = extractBooleanParam(
-    searchParams.get("shouldWaitForResponse"),
-    false
-  );
+  const {
+    deliveryInfoDecoded,
+    sendDeliveryInfoThroughXMTP,
+    shouldWaitForResponse,
+    postDeliveryInfoHeadersDecoded,
+    postDeliveryInfoUrl,
+    postRedemptionConfirmedHeadersDecoded,
+    postRedemptionConfirmedUrl,
+    postRedemptionSubmittedHeadersDecoded,
+    postRedemptionSubmittedUrl,
+    targetOrigin,
+    eventTag
+  } = parseDeliveryInfoData(searchParams);
 
-  const postDeliveryInfoUrl =
-    searchParams.get("postDeliveryInfoUrl") || undefined;
-  const postDeliveryInfoHeaders =
-    searchParams.get("postDeliveryInfoHeaders") || undefined;
-  let postDeliveryInfoHeadersDecoded = undefined;
-  if (postDeliveryInfoHeaders) {
-    try {
-      postDeliveryInfoHeadersDecoded = JSON.parse(postDeliveryInfoHeaders);
-    } catch (e) {
-      console.error(
-        `Unable to parse JSON from postDeliveryInfoHeaders='${postDeliveryInfoHeaders}': ${e}`
-      );
-    }
-  }
-  const postRedemptionSubmittedUrl =
-    searchParams.get("postRedemptionSubmittedUrl") || undefined;
-  const postRedemptionSubmittedHeaders =
-    searchParams.get("postRedemptionSubmittedHeaders") || undefined;
-  let postRedemptionSubmittedHeadersDecoded = undefined;
-  if (postRedemptionSubmittedHeaders) {
-    try {
-      postRedemptionSubmittedHeadersDecoded = JSON.parse(
-        postRedemptionSubmittedHeaders
-      );
-    } catch (e) {
-      console.error(
-        `Unable to parse JSON from postRedemptionSubmittedHeaders='${postRedemptionSubmittedHeaders}': ${e}`
-      );
-    }
-  }
-  const postRedemptionConfirmedUrl =
-    searchParams.get("postRedemptionConfirmedUrl") || undefined;
-  const postRedemptionConfirmedHeaders =
-    searchParams.get("postRedemptionConfirmedHeaders") || undefined;
-  let postRedemptionConfirmedHeadersDecoded = undefined;
-  if (postRedemptionConfirmedHeaders) {
-    try {
-      postRedemptionConfirmedHeadersDecoded = JSON.parse(
-        postRedemptionConfirmedHeaders
-      );
-    } catch (e) {
-      console.error(
-        `Unable to parse JSON from postRedemptionConfirmedHeaders='${postRedemptionConfirmedHeaders}': ${e}`
-      );
-    }
-  }
   const configId = searchParams.get("configId") as ConfigId;
   if (!configId) {
     return <p>Missing 'configId' query param</p>;
@@ -109,7 +59,6 @@ export function Redeem() {
   const signaturesStr = searchParams.get("signatures") as string;
   const signatures = signaturesStr ? signaturesStr.split(",") : undefined;
 
-  const eventTag = searchParams.get("eventTag") as string;
   const lookAndFeel =
     (searchParams.get("lookAndFeel") as "regular" | "modal") || "regular";
   // In case the deliveryInfo shall be transferred between frontend windows, the targetOrigin
@@ -121,7 +70,7 @@ export function Redeem() {
     <RedemptionWidget
       withReduxProvider={true}
       withCustomReduxContext={false}
-      withWeb3React={false}
+      withWeb3React={true}
       withExternalSigner={withExternalSigner === "true"}
       showRedemptionOverview={showRedemptionOverview}
       sendDeliveryInfoThroughXMTP={sendDeliveryInfoThroughXMTP}
@@ -150,7 +99,6 @@ export function Redeem() {
       ipfsGateway={CONFIG.ipfsGateway as string}
       ipfsProjectId={CONFIG.ipfsProjectId}
       ipfsProjectSecret={CONFIG.ipfsProjectSecret}
-      children={<></>}
       walletConnectProjectId={CONFIG.walletConnectProjectId as string}
       fairExchangePolicyRules={CONFIG.fairExchangePolicyRules as string}
       raiseDisputeForExchangeUrl={CONFIG.raiseDisputeForExchange as string}
@@ -161,107 +109,17 @@ export function Redeem() {
           console.error(`Unable to post message ${e}`);
         }
       }}
-      deliveryInfoHandler={
+      deliveryInfoHandler={createDeliveryInfoHandler(
+        targetOrigin,
+        shouldWaitForResponse,
+        eventTag
+      )}
+      redemptionSubmittedHandler={createRedemptionSubmittedHandler(
         targetOrigin
-          ? async (message, signature) => {
-              try {
-                const event = {
-                  type: "boson-delivery-info",
-                  message,
-                  signature,
-                  tag: eventTag
-                };
-                // precaution: register to the response before posting the message
-                const responseType = "boson-delivery-info-response";
-                const _waitForResponse = shouldWaitForResponse
-                  ? (waitForResponse(responseType) as Promise<{
-                      message: DeliveryInfoCallbackResponse;
-                      origin: string;
-                    }>)
-                  : undefined;
-                // post the message
-                console.log(
-                  `Post '${event.type}' message to '${targetOrigin}'`
-                );
-                window.parent.postMessage(event, targetOrigin);
-                if (_waitForResponse) {
-                  console.log(`Wait for response '${responseType}' message`);
-                  const response = await _waitForResponse;
-                  console.log(
-                    `Received response '${responseType}' from '${
-                      response.origin
-                    }'. Content: '${JSON.stringify(
-                      JSON.stringify(response.message)
-                    )}'`
-                  );
-                  return response.message;
-                }
-                return {
-                  accepted: true,
-                  resume: true,
-                  reason: ""
-                };
-              } catch (e) {
-                console.error(`Unable to post message ${e}`);
-                return {
-                  accepted: false,
-                  reason: "",
-                  resume: false
-                };
-              }
-            }
-          : undefined
-      }
-      redemptionSubmittedHandler={
+      )}
+      redemptionConfirmedHandler={createRedemptionConfirmedHandler(
         targetOrigin
-          ? async (message) => {
-              try {
-                const event = {
-                  type: "boson-redemption-submitted",
-                  message
-                };
-                // post the message
-                console.log(`Post ${event.type} message to ${targetOrigin}`);
-                window.parent.postMessage(event, targetOrigin);
-                return {
-                  accepted: true,
-                  reason: ""
-                };
-              } catch (e) {
-                console.error(`Unable to post message ${e}`);
-                return {
-                  accepted: false,
-                  reason: ""
-                };
-              }
-            }
-          : undefined
-      }
-      redemptionConfirmedHandler={
-        targetOrigin
-          ? async (message) => {
-              try {
-                const event = {
-                  type: "boson-redemption-confirmed",
-                  message
-                };
-                // post the message
-                console.log(`Post ${event.type} message to ${targetOrigin}`);
-                window.parent.postMessage(event, targetOrigin);
-                return {
-                  accepted: true,
-                  reason: ""
-                };
-              } catch (e) {
-                console.error(`Unable to post message ${e}`);
-                return {
-                  accepted: false,
-                  reason: ""
-                };
-              }
-            }
-          : undefined
-      }
+      )}
       modalMargin="2%"
       widgetAction={widgetAction}
       deliveryInfo={deliveryInfoDecoded}
@@ -318,33 +176,4 @@ function checkExchangeState(
     }
   }
   throw new Error(`Not supported exchange state '${exchangeStateStr}'`);
-}
-
-function extractBooleanParam(
-  paramStr: string | null,
-  defaultValue: boolean
-): boolean {
-  return paramStr ? /^true$/i.test(paramStr) : defaultValue;
-}
-
-async function waitForResponse(
-  response: string
-): Promise<{ message: unknown; origin: string }> {
-  return new Promise<{ message: unknown; origin: string }>(
-    (resolve, reject) => {
-      const eventType = "message";
-      const listener = (event: MessageEvent | undefined) => {
-        try {
-          if (event?.data?.type === response) {
-            // Ensure the listener won't be called again. Do not use "once" option because some other message could be received in the meanwhile
-            window.removeEventListener(eventType, listener);
-            resolve({ message: event.data.message, origin: event.origin });
-          }
-        } catch (e) {
-          reject(e);
-        }
-      };
-      window.addEventListener(eventType, listener);
-    }
-  );
 }
