+++ src/utils/params.ts
@@ -1,0 +1,6 @@
+export function extractBooleanParam(
+  paramStr: string | null,
+  defaultValue: boolean
+): boolean {
+  return paramStr ? /^true$/i.test(paramStr) : defaultValue;
+}
